requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
beanie==1.27.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
coverage==7.6.4
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jose==1.0.0
lazy-model==0.2.0
motor==3.6.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.32.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
beanie==1.27.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
coverage==7.6.4
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jose==1.0.0
lazy-model==0.2.0
motor==3.6.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.32.0
